% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenLight_creator.R
\name{genlight_creator}
\alias{genlight_creator}
\title{Creates a genlight object from a vcf file that can be used in adegent package to look at population structure}
\usage{
genlight_creator(vcf_file = "c_ven_phy_50_pop.vcf", pop_data = "c_ven_pop_5.txt", genlight_name = "gl.test", sample_col = sample, pop_col = sub_species)
}
\arguments{
\item{vcf_file}{the path to the vcf file that will be used to create the genlight object}

\item{pop_data}{a text file with the samples and population data to be analyzed, should have a heading for each of the columns}

\item{genlight_name}{the name of the output genelight object, needs to be in quotes}

\item{sample_col}{the name of the column that contains the sample names}

\item{pop_col}{the column from the population text file to be used as the population data}
}
\description{
need packages ape, poppr, vcfR
}
