% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpPopData.R
\name{pop_data_output}
\alias{pop_data_output}
\title{Produce a .txt file with needed information for population data}
\usage{
pop_data_output(file = "/cloud/project/data/DNA_Samples_Data_Analysis.csv", columns = c(SJ_number, Species, sub_species), sample_col = SJ_number, samples = c("AR01", "AR02", "AR03"), NA_values = c("", "na"), output_file = "/cloud/project/Output_files/test_pop_data.txt")
}
\arguments{
\item{file}{the file path to the csv file to be converted to text file}

\item{columns}{the columns in the csv file that are to be kept for the text file}

\item{sample_col}{the column that has the sample numbers/names that will be used to select certain rows}

\item{samples}{the samples that are to be kept for the text file}

\item{NA_values}{values in the csv file that should be read as NA values}

\item{output_file}{the file path for the output text file}
}
\description{
need tidyverse package
}
