% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sample_counts.R
\name{samp_2B_seq}
\alias{samp_2B_seq}
\title{Use to check number of samples with a certain variable and then filter out samples that have a certain value for that variable and create a new csv file with those samples}
\usage{
samp_2B_seq(file = "/cloud/project/data/DNA_Samples_Data_Analysis.csv", na_values = c("", "na"), col2count = strange_topology, output_csv = "/cloud/project/Output_files/test_output.csv", filter = c("yes", NA))
}
\arguments{
\item{file}{a csv file with the sample data}

\item{na_values}{list of values in csv file to be treated as NA values}

\item{col2count}{the column in the csv file to use as the variable to see how many sample belong to each variable}

\item{output_csv}{the file path and name of the csv file with the filtered data}

\item{filter}{list of values for the selected column that is to be filtered into a new csv file}
}
\description{
need tidyverse package
}
