% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure_plot.R
\name{snmf_structure_plot}
\alias{snmf_structure_plot}
\title{need LEA package}
\usage{
snmf_structure_plot(project_name = snmf_test_2, K = 4, colors = c("green", "blue", "purple", "red"))
}
\arguments{
\item{project_name}{the project name that was created through snmf and using the SNMF_Analysis function. Can also be any snmf Project}

\item{K}{the number of populations that are to be used for the structure analysis}

\item{colors}{a list of colors to be used for the structure plot, need the same amount of colors as K}
}
\description{
need LEA package
}
